\name{TetraQTLAnalysis-package}
\alias{TetraQTLAnalysis-package}
\alias{TetraQTLAnalysis}
\docType{package}
\title{
\packageTitle{TetraQTLAnalysis}
}
\description{
\packageDescription{TetraQTLAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TetraQTLAnalysis}
\packageIndices{TetraQTLAnalysis}
A package implemented for QTL analysis in autotetraploid species, which includes five fuctions (TraitSeg,Heritability,OrthogonalModel,BvMethod and QvMethod) and four datasets(LinkageMap,SimulatedDatasetB,SimulatedDatasetQ,TraitData) for demonstration purpose. 
}
\author{
\packageAuthor{TetraQTLAnalysis}

Maintainer: \packageMaintainer{TetraQTLAnalysis}
}
\references{
Chen J, et al. 2018 Orthogonal contrast based models for quantitative genetic analysis in autotetraploid species. New Phytologist 220:332-346. 
Chen J, et al. 2019 Mapping Quantitative Trait Loci in Autotetraploid Species.(Under review)
}